feat: implementa ServicoLocalItemService

+++ src/main/java/com/gestaooserp/dev/repository/ServicoLocalItemRepository.java
@@ -3,5 +3,5 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.ServicoLocalItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServicoLocalItemRepository extends JpaRepository<ServicoLocalItem,Integer> {
+public interface ServicoLocalItemRepository extends JpaRepository<ServicoLocalItem,Long> {
 }
